clean up scripts a bit

+++ scripts/clean.sh
@@ -1,9 +1,6 @@
 #!/bin/bash
 source .zefie/scripts/buildenv.sh
-if [ -z "${KERNEL_BUILDDIR}" ]; then
-	exit 1; # die before we destroy stuff
-fi
-rm -rf "${KERNEL_BUILDDIR}"/*
+rm -rf "${KERNEL_BUILDDIR:?}"/*
 .zefie/scripts/resetgit.sh
 RC=$?
 if [ $RC -ne 0 ]; then
@@ -16,7 +13,7 @@ if [ $RC -ne 0 ]; then
 	echo "Error doing kernel mrproper."
 	exit $RC
 fi
-rm -f "${DEFCONFIG_DIR}/"${KERNEL_NAME_LOWER}_*_defconfig
+rm -f "${DEFCONFIG_DIR}/${KERNEL_NAME_LOWER}_"*_defconfig
 .zefie/scripts/defconfig.sh
 RC=$?
 if [ $RC -ne 0 ]; then
